Moves alert box to it's own file
